+++ apx_core/src/http_signatures/verify.rs
@@ -98,6 +98,8 @@ pub struct HttpSignatureData {
     pub base: String, // recreated signature base
     pub signature: Vec<u8>,
     pub expires_at: DateTime<Utc>,
+    /// The authority of the target URI
+    pub authority: String,
     pub content_digest: Option<ContentDigest>,
 }
 
@@ -190,6 +192,11 @@ pub fn parse_http_signature_cavage(
     let signature_b64 = signature_parameters.get("signature")
         .ok_or(VerificationError::ParseError("signature is missing"))?;
     let signature = base64::decode(signature_b64)?;
+    let target_authority = request_headers.get("Host")
+        .ok_or(VerificationError::header_missing("Host"))?
+        .to_str()
+        .map_err(|_| VerificationError::header_value("Host"))?
+        .to_string();
     let created_at = if let Some(created_at) = signature_parameters.get("created") {
         let create_at_timestamp = created_at.parse()
             .map_err(|_| VerificationError::ParseError("invalid timestamp"))?;
@@ -257,6 +264,7 @@ pub fn parse_http_signature_cavage(
         base: signature_base,
         signature,
         expires_at,
+        authority: target_authority,
         content_digest: maybe_digest,
         is_rfc9421: false,
     };
@@ -398,6 +406,7 @@ pub fn parse_http_signature_rfc9421(
         base: signature_base,
         signature: signature_value,
         expires_at,
+        authority: target_authority,
         content_digest: maybe_digest,
         is_rfc9421: true,
     };